feat: use clones-with-immutable-args for QTokens
also refactor and fix order of parameters on oracles and the PriceRegistry

+++ contracts/interfaces/IChainlinkFixedTimeOracleManager.sol
@@ -5,11 +5,11 @@ import "./IChainlinkOracleManager.sol";
 
 interface IChainlinkFixedTimeOracleManager is IChainlinkOracleManager {
     /// @notice emitted when a new time is added for fixed updates
-    event FixedTimeUpdate(uint256 fixedTime, bool isValidTime);
+    event FixedTimeUpdate(uint24 fixedTime, bool isValidTime);
 
     /// @notice Validate or invalidated a given fixed time for updates
-    function setFixedTimeUpdate(uint256 fixedTime, bool isValidTime) external;
+    function setFixedTimeUpdate(uint24 fixedTime, bool isValidTime) external;
 
     /// @notice fixed time => is allowed
-    function chainlinkFixedTimeUpdates(uint256) external view returns (bool);
+    function chainlinkFixedTimeUpdates(uint24) external view returns (bool);
 }
